s9: создан отдельный класс для вывода историии и приоритета
